Separated courant number diagnostic core logic from SWE operator.
This PR pulls some basic logic for our courant number diagnostics
(including a MPI datatype and operator for reducing maximum wave
 speeds) out of the SWE code. The actual determination of the max
wave speed remains part of the SWE code, but as we generalize our
approach to "operators", this will also be pulled out.
